fix: address review — merge file-utils into bounded-fs, scan TS-path skill installs, router-shaped probe globs
IGNORED_DIRS gains site-packages/DerivedData/Pods and is the single
source for walk skips and glob ignores. downloadSkill scans every
install (orchestrator cache and reference included) with the same
poisoned-skill scan as the Bash hook, removing the install on a match.
TanStack/React Router content probes glob route-shaped files instead
of every source file.

Generated-By: PostHog Code
Task-Id: 64dfb953-a7c2-44d2-a241-89f06bf7e4f7
