+++ src/steps/add-or-update-environment-variables.ts
@@ -2,7 +2,7 @@ import type { Integration } from '@lib/constants';
 import { withProgress } from '../telemetry';
 import { analytics } from '@utils/analytics';
 import { getUI } from '@ui';
-import { getDotGitignore } from '@utils/file-utils';
+import { getDotGitignore } from '@utils/bounded-fs';
 import * as fs from 'fs';
 import path from 'path';
